Initial commit for a tool that will create a dataset for self-supervised learning.
image_cutter detects the moving ball on the original image of a pool table using computer vision techniques.
It  "cuts out" images of the ball out of input video, and saves them into a binary proto file. They will be
pasted over the image of the background, to create the synthetic training set for training the neural network.

training_set_generator loads images of individual balls, loads the image of the background, and generates
images for t
git commit